when sampling molecules (without saving the chain of intermediate states), pad batch only up to size of largest molecule
When sampling from a trained model, the function `qm9.sampling.sample_different_sizes_and_save()` will pad a batch
to the size of the largest molecule in the training dataset. There are a few issues with this. First, this means that
unnecessary computations are done when the largest molecule in the batch is smaller than the largest molecule in the training
set. I'm sure the authors will be aware of this fact, since the `Dataloader` for training only pads up to the largest molecule in eatch batch.

Second, this makes it impossible to sample molecules that are larger than any seen in the training dataset, even though
the model is theoretically capable of sampling molecules of any size. It would useful to understand if and how well this model
extrapolates to molecules of sizes it has not seen before.