pad batches of sampled molecules when analyzing model during training
the function train_test.analyze_and_save() make multiple calls to the sampling function, and then
concatenates the result. In a previous commit, I changed the sampling function to pad molecules up
to the largest size in the batch. This meant that in the analyze_and_save() function, that batches
of sampled molecules had different shapes (along the dimension corresponding to the number of nodes)
and therefore this function would throw an error when trying to concatenate tensors from multiple
calls to the sampling function. this commit pads the batches of sampled molecules so that they
can be succesfully concatenated.